+++ src/main/kotlin/org/antop/board/common/extensions/Authentication.kt
@@ -6,3 +6,9 @@ val Authentication.isAnonymous: Boolean
     get() {
         return principal == "anonymousUser"
     }
+
+/**
+ * 인증 되었고 익명 유저도 아님
+ */
+val Authentication.isRealAuthenticated: Boolean
+    get() = isAuthenticated && !isAnonymous
