+++ src/main/kotlin/org/antop/board/member/repository/MemberRepository.kt
@@ -1,17 +1,15 @@
 package org.antop.board.member.repository
 
 import org.antop.board.member.model.Member
-import org.antop.board.member.model.Members
-import org.springframework.stereotype.Repository
-import org.springframework.transaction.annotation.Transactional
 
-@Repository
-@Transactional(readOnly = true)
-class MemberRepository {
-    fun findByEmail(email: String) =
-        Member
-            .find { Members.email eq email }
-            .firstOrNull()
+interface MemberRepository {
+    /**
+     * [id]로 회원 조회
+     */
+    fun findById(id: Long): Member?
 
-    fun findById(id: Long) = Member.findById(id)
+    /**
+     * [email]로 회원 조회
+     */
+    fun findByEmail(email: String): Member?
 }
