+++ src/main/kotlin/org/antop/board/post/model/Comments.kt
@@ -1,10 +1,11 @@
 package org.antop.board.post.model
 
+import org.antop.board.member.model.Members
 import org.jetbrains.exposed.dao.id.LongIdTable
 import org.jetbrains.exposed.sql.kotlin.datetime.datetime
 
 object Comments : LongIdTable(name = "comments", columnName = "comment_id") {
-    val authorId = long("author_id")
+    val authorId = long("author_id").references(Members.id)
     val content = text("content")
     val created = datetime("created")
     val modified = datetime("modified").nullable()
