+++ src/main/kotlin/org/antop/board/post/model/Posts.kt
@@ -1,13 +1,14 @@
 package org.antop.board.post.model
 
 import org.antop.board.common.exposed.tags
+import org.antop.board.member.model.Members
 import org.jetbrains.exposed.dao.id.LongIdTable
 import org.jetbrains.exposed.sql.kotlin.datetime.datetime
 
 object Posts : LongIdTable(name = "posts", columnName = "post_id") {
     val subject = varchar("subject", 255)
     val content = largeText("content")
-    val authorId = long("author_id")
+    val authorId = long("author_id").references(Members.id)
     val created = datetime("created")
     val modified = datetime("modified").nullable()
     val tags = tags("tags").nullable()
